requirements.txt
<<<<<<< HEAD
# Core dependencies
apify-client>=2.3.0
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.3.0
pandas==2.2.2

# ML/AI models (CPU-optimized)
torch==2.3.0
torchvision==0.18.0
transformers==4.41.0
sentence-transformers==2.7.0
detoxify==0.5.2
openai-whisper==20231117
NudeNet>=3.4.0

# Media processing
opencv-python==4.9.0.80
ffmpeg-python==0.2.0

# Utilities
tqdm==4.66.4
aiohttp==3.9.5
scikit-learn==1.5.0
numpy==1.26.4
pyarrow==15.0.2

# Optional but recommended
fastparquet==2024.2.0
=======
# Core dependencies
apify-client>=2.3.0
python-dotenv==1.0.1
requests==2.31.0
Pillow==10.3.0
pandas==2.2.2

# ML/AI models (CPU-optimized)
torch==2.3.0
torchvision==0.18.0
transformers==4.41.0
sentence-transformers==2.7.0
detoxify==0.5.2
openai-whisper==20231117
NudeNet>=3.4.0

# Media processing
opencv-python==4.9.0.80
ffmpeg-python==0.2.0

# Utilities
tqdm==4.66.4
aiohttp==3.9.5
scikit-learn==1.5.0
numpy==1.26.4
pyarrow==15.0.2

# Optional but recommended
fastparquet==2024.2.0
>>>>>>> 8bf316f0f1e5a33d7051e28b6dde11854481b5fc
